+++ cpp/serve/engine_actions/batch_prefill_base.cc
@@ -130,7 +130,8 @@ bool BatchPrefillBaseActionObj::CanPrefill(EngineState estate, int num_prefill_r
                         ? (engine_config_->spec_draft_length + 1)
                         : 1;
   if ((num_running_rsentries + num_prefill_rsentries) * spec_factor >
-      std::min(engine_config_->max_num_sequence, engine_config_->prefill_chunk_size)) {
+      std::min(static_cast<int64_t>(engine_config_->max_num_sequence),
+               engine_config_->prefill_chunk_size)) {
     return false;
   }
